% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames_raw.R
\docType{data}
\name{ames_schools}
\alias{ames_schools}
\title{Ames Public Schools}
\value{
\item{ames_schools}{a tibble}
}
\description{
Locations of local schools
}
\details{
The data set includes the school name and the geocodes.
}
\keyword{datasets}
