% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames_raw.R
\docType{data}
\name{ames_raw}
\alias{ames_raw}
\alias{ames_geo}
\title{Raw Ames Housing Data}
\source{
De Cock, D. (2011). "Ames, Iowa: Alternative to the Boston Housing Data as an End of Semester Regression Project," \emph{Journal of Statistics Education},  Volume 19, Number 3.

\url{https://ww2.amstat.org/publications/jse/v19n3/decock/DataDocumentation.txt}

\url{http://ww2.amstat.org/publications/jse/v19n3/decock.pdf}
}
\value{
\item{ames_raw}{a tibble}
}
\description{
Summon the data described by De Cock (2011) where 82 fields were recored for 2,930 properties in Ames IA.
}
\keyword{datasets}
