% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ames.R
\name{make_ames}
\alias{make_ames}
\title{Create a Processed Version of the Ames Housing Data}
\usage{
make_ames()
}
\value{
A tibble with the data without ordered factors
}
\description{
Create a Processed Version of the Ames Housing Data
}
